// src/types/generated/Abi.ts
export declare namespace NFTMarketplace {
  export type ListedTokenStruct = {
    tokenId: BigNumberish;
    owner: AddressLike;
    seller: AddressLike;
    price: BigNumberish;
    curentlyListed: boolean;
  };

  export type ListedTokenStructOutput = [
    tokenId: bigint,
    owner: string,
    seller: string,
    price: bigint,
    curentlyListed: boolean
  ] & {
    tokenId: bigint;
    owner: string;
    seller: string;
    price: bigint;
    curentlyListed: boolean;
  };
}